moving out TileRecordCodec to a class
